+++ CHANGELOG.md
@@ -7,7 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Add support for `x**y` with power-of-two `x` or constant `y` (#193)
 - Add limited support for `$countones`, `$onehot`, `$onehot0`, `$isunknown`, `$left`, `$right`, `$low`, `$high`, `$increment`, `$size` (#204)
-- Add proper emission of string literals
+- Add emission of string literals (#205)
+- Add constant casts between `string` and bit vectors (#205)
 
 ### Changed
 - Make unsupported system task an error (#206)